+++ dcc-portal-server/src/main/java/org/icgc/dcc/portal/server/resource/core/ManifestResource.java
@@ -18,6 +18,7 @@
 package org.icgc.dcc.portal.server.resource.core;
 
 import static com.google.common.base.Strings.isNullOrEmpty;
+import static com.google.common.net.HttpHeaders.CACHE_CONTROL;
 import static com.google.common.net.HttpHeaders.CONTENT_DISPOSITION;
 import static com.sun.jersey.multipart.Boundary.BOUNDARY_PARAMETER;
 import static com.sun.jersey.multipart.MultiPartMediaTypes.MULTIPART_MIXED;
@@ -25,18 +26,23 @@ import static com.sun.jersey.multipart.MultiPartMediaTypes.MULTIPART_MIXED_TYPE;
 import static java.util.Collections.singletonMap;
 import static java.util.stream.Collectors.toList;
 import static javax.ws.rs.core.MediaType.APPLICATION_FORM_URLENCODED;
-import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
+import static javax.ws.rs.core.MediaType.APPLICATION_JSON_TYPE;
 import static javax.ws.rs.core.MediaType.TEXT_PLAIN;
 import static javax.ws.rs.core.MediaType.TEXT_PLAIN_TYPE;
-import static javax.ws.rs.core.Response.ok;
-import static org.icgc.dcc.portal.server.util.MediaTypes.GZIP;
+import static javax.ws.rs.core.Response.Status.OK;
+import static org.icgc.dcc.portal.server.manifest.model.ManifestFormat.FILES;
+import static org.icgc.dcc.portal.server.manifest.model.ManifestFormat.JSON;
+import static org.icgc.dcc.portal.server.manifest.model.ManifestFormat.TARBALL;
+import static org.icgc.dcc.portal.server.util.MediaTypes.GZIP_TYPE;
 
+import java.io.IOException;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.Date;
 import java.util.List;
 import java.util.UUID;
 
+import javax.servlet.http.HttpServletResponse;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DefaultValue;
 import javax.ws.rs.FormParam;
@@ -46,9 +52,8 @@ import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
+import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.StreamingOutput;
 
 import org.icgc.dcc.portal.server.manifest.ManifestContext;
 import org.icgc.dcc.portal.server.manifest.ManifestService;
@@ -69,17 +74,18 @@ import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.SwaggerDefinition;
 import io.swagger.annotations.Tag;
-import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
+import lombok.SneakyThrows;
 import lombok.val;
 import lombok.extern.slf4j.Slf4j;
 
+/**
+ * Resource responsible for serving and storing representations of native download manifests.
+ */
 @Slf4j
 @Component
 @Path("/v1/manifests")
 @Api(tags = "manifests")
 @SwaggerDefinition(tags = @Tag(name = "manifests", description = "Resources about manifests"))
-@RequiredArgsConstructor(onConstructor = @__({ @Autowired }))
 public class ManifestResource extends Resource {
 
   /**
@@ -126,13 +132,19 @@ public class ManifestResource extends Resource {
   /**
    * Dependencies.
    */
-  @NonNull
-  private final ManifestService manifestService;
+  @Autowired
+  private ManifestService manifestService;
+
+  /**
+   * State.
+   */
+  @Context
+  private HttpServletResponse response;
 
   @GET
   @Timed
   @ApiOperation(value = "Generate manifest(s) on the fly as specified by the supplied parameters.", notes = "Does not store the result")
-  public Response generateManifest(
+  public void generateManifest(
       // Input
       @ApiParam(value = REPOS_DESC, allowMultiple = true) @QueryParam(REPOS_PARAM) @DefaultValue("") ListParam repoParam,
       @ApiParam(value = FILTERS_DESC) @QueryParam(FILTERS_PARAM) @DefaultValue(FILTERS_DEFAULT) FiltersParam filtersParam,
@@ -145,7 +157,7 @@ public class ManifestResource extends Resource {
     val manifest = createManifest(repoParam, filtersParam, unique, format, fieldsParam, multipart);
 
     log.debug("Rendering manifest: {}", manifest);
-    return renderManifest(manifest);
+    renderManifest(manifest);
   }
 
   @POST
@@ -174,7 +186,7 @@ public class ManifestResource extends Resource {
   @GET
   @Path("/{" + MANIFEST_ID_PARAM + "}")
   @ApiOperation(value = "Retrieves a manifest by its ID, overriding output specification as indicated", notes = "Everything but the `filter` may be overriden")
-  public Response getManifest(
+  public void getManifest(
       // Input
       @ApiParam(value = MANIFEST_ID_DESC, required = true) @PathParam(MANIFEST_ID_PARAM) UUID manifestId,
 
@@ -219,7 +231,7 @@ public class ManifestResource extends Resource {
     validateManifest(manifest);
 
     log.debug("Rendering manifest: {}", manifest);
-    return renderManifest(manifest);
+    renderManifest(manifest);
   }
 
   private Manifest createManifest(
@@ -243,6 +255,7 @@ public class ManifestResource extends Resource {
         .setMultipart(multipart);
 
     validateManifest(manifest);
+
     return manifest;
   }
 
@@ -253,58 +266,58 @@ public class ManifestResource extends Resource {
         "If `multipart` is specified then format must be `files`");
   }
 
-  private Response renderManifest(Manifest manifest) {
-    switch (manifest.getFormat()) {
-    case TARBALL:
-      return tarball(manifest);
-    case FILES:
-      return files(manifest);
-    case JSON:
-      return json(manifest);
-    default:
-      throw new IllegalStateException();
-    }
-  }
-
-  private Response tarball(Manifest manifest) {
-    return ok(entityStream(manifest), GZIP)
-        .header(CONTENT_DISPOSITION, attachmentContent(manifest))
-        .cacheControl(noCache())
-        .build();
-  }
-
-  private Response files(Manifest manifest) {
-    if (manifest.isMultipart()) {
-      val mediaType = multipartMixed(manifest.getTimestamp());
-      return ok(entityStream(manifest), mediaType)
-          .cacheControl(noCache())
-          .build();
+  private void renderManifest(Manifest manifest) {
+    val format = manifest.getFormat();
+    if (format == TARBALL) {
+      renderTarball(manifest);
+    } else if (format == FILES) {
+      renderFiles(manifest);
+    } else if (format == JSON) {
+      renderJson(manifest);
     } else {
-      return ok(entityStream(manifest), TEXT_PLAIN_TYPE)
-          .header(CONTENT_DISPOSITION, attachmentContent(manifest))
-          .cacheControl(noCache())
-          .build();
+      checkFormat(format.getKey());
     }
   }
 
-  private Response json(Manifest manifest) {
-    return ok(entityStream(manifest), APPLICATION_JSON).build();
+  @SneakyThrows
+  private void renderTarball(Manifest manifest) {
+    render(manifest, GZIP_TYPE, true);
   }
 
-  private StreamingOutput entityStream(Manifest manifest) {
-    // Stream output
-    return output -> manifestService.generateManifests(new ManifestContext(manifest, output));
+  private void renderFiles(Manifest manifest) {
+    if (manifest.isMultipart()) {
+      render(manifest, multipartMixed(manifest.getTimestamp()), false);
+    } else {
+      render(manifest, TEXT_PLAIN_TYPE, true);
+    }
   }
 
-  private static List<ManifestField> parseFields(ListParam fieldsParam) {
-    return fieldsParam.get().stream().map(ManifestField::fromString).collect(toList());
+  private void renderJson(Manifest manifest) {
+    render(manifest, APPLICATION_JSON_TYPE, false);
   }
 
-  private static void checkFormat(String format) {
-    checkRequest(isNullOrEmpty(format),
-        "`format` is required");
-    checkRequest(!isValidEnum(ManifestFormat.class, format.toUpperCase()),
-        "`format` is not a valid value in: %s", Arrays.toString(ManifestFormat.values()).toLowerCase());
+  @SneakyThrows
+  private void render(Manifest manifest, MediaType mediaType, boolean attachment) {
+    // Using {@link HttpServletResponse} is required to commit the response code and headers eagerly. Typically one
+    // would use {@link StreamingOutput} but this commits headers on the first write. Since it takes a long time to
+    // issue the first write for a manifest containing a large number of files, this causes considerable delays for a
+    // user agent to display the download dialog.
+    response.setStatus(OK.getStatusCode());
+    response.setContentType(mediaType.toString());
+    response.addHeader(CACHE_CONTROL, noCache().toString());
+    if (attachment) response.addHeader(CONTENT_DISPOSITION, attachmentContent(manifest).toString());
+
+    // Calling this method automatically commits the response, meaning the status code and headers will be written.
+    response.flushBuffer();
+
+    try {
+      // Write the manifests
+      val context = new ManifestContext(manifest, response.getOutputStream());
+      manifestService.generateManifests(context);
+    } catch (IOException e) {
+      // This can happen when a user aborts the connection while streaming
+      log.warn("Error generating manifests: {}", e.getMessage());
+    }
   }
 
   private ContentDisposition attachmentContent(Manifest manifest) {
@@ -321,6 +334,17 @@ public class ManifestResource extends Resource {
     return new MediaType(
         MULTIPART_MIXED_TYPE.getType(), MULTIPART_MIXED_TYPE.getSubtype(),
         singletonMap(BOUNDARY_PARAMETER, "boundary_" + timestamp));
+  }
+
+  private static List<ManifestField> parseFields(ListParam fieldsParam) {
+    return fieldsParam.get().stream().map(ManifestField::fromString).collect(toList());
+  }
+
+  private static void checkFormat(String format) {
+    checkRequest(isNullOrEmpty(format),
+        "`format` is required");
+    checkRequest(!isValidEnum(ManifestFormat.class, format.toUpperCase()),
+        "`format` is not a valid value in: %s", Arrays.toString(ManifestFormat.values()).toLowerCase());
   }
 
 }