+++ dcc-portal-server/src/test/java/org/icgc/dcc/portal/server/repository/DonorRepositoryTest.java
@@ -17,29 +17,24 @@
 
 package org.icgc.dcc.portal.server.repository;
 
-import static com.google.common.collect.Lists.newArrayList;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.icgc.dcc.portal.server.model.IndexModel.FIELDS_MAPPING;
 import static org.icgc.dcc.portal.server.util.ElasticsearchResponseUtils.getString;
-import static org.junit.Assert.fail;
 
 import java.util.Map;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 import org.dcc.portal.pql.query.QueryEngine;
 import org.elasticsearch.action.search.SearchResponse;
 import org.elasticsearch.search.SearchHit;
 import org.elasticsearch.search.SearchHits;
-import org.icgc.dcc.portal.server.config.ServerProperties;
 import org.icgc.dcc.portal.server.model.BaseEntitySet;
 import org.icgc.dcc.portal.server.model.EntitySet;
 import org.icgc.dcc.portal.server.model.EntitySet.State;
 import org.icgc.dcc.portal.server.model.IndexModel.Kind;
 import org.icgc.dcc.portal.server.model.IndexModel.Type;
-import org.icgc.dcc.portal.server.model.param.FiltersParam;
 import org.icgc.dcc.portal.server.model.Query;
-import org.icgc.dcc.portal.server.repository.TermsLookupRepository.TermLookupType;
+import org.icgc.dcc.portal.server.model.param.FiltersParam;
 import org.icgc.dcc.portal.server.test.TestIndex;
 import org.junit.Before;
 import org.junit.Test;
@@ -275,59 +270,9 @@ public class DonorRepositoryTest extends BaseElasticSearchTest {
     donorRepository.findAll(query);
   }
 
-  @Test
-  public void testPhenotypeAnalysis() {
-    val id1 = UUID.randomUUID();
-    val id2 = UUID.randomUUID();
-    setUpTermsLookup(id1, id2);
-
-    val results = donorRepository.getPhenotypeAnalysisResult(newArrayList(id1, id2));
-
-    // Here we are only interested in results that contain stats, which is under "ageAtDiagnosisGroup", and we use
-    // flatMap to flatten/unwrap the result to a simple list.
-    val data = results.stream()
-        .filter(result -> result.getName().equals("ageAtDiagnosisGroup"))
-        .flatMap(result -> result.getData().stream())
-        .collect(Collectors.toList());
-
-    assertThat(data.size()).isEqualTo(2);
-
-    for (val analysisResult : data) {
-      val uuid = analysisResult.getId();
-      val mean = analysisResult.getSummary().getMean();
-
-      // Note: the expected mean values used below depend on the fixture data defined in
-      // /dcc-portal-server/src/test/resources/fixtures/DonorRepositoryTest.json. If data in that file has changed, these
-      // values might no longer represent the expected result. In that case, update these accordingly.
-      if (uuid.equals(id1)) {
-        assertThat(mean).isEqualTo(109.6);
-      } else if (uuid.equals(id2)) {
-        assertThat(mean).isEqualTo(135.66666666666666);
-      } else {
-        // We shouldn't be here but, if this happens, we fail this test.
-        fail("Encountered an unexpected UUID in query result.");
-      }
-
-    }
-
-  }
-
   @Override
   protected Object cast(Object object) {
     return object;
-  }
-
-  private void setUpTermsLookup(final UUID id1, final UUID id2) {
-    val termsLookupRepository =
-        new TermsLookupRepository(es.client(), TestIndex.RELEASE.getName(), TestIndex.REPOSITORY.getName(),
-            new ServerProperties());
-    val lookupType = TermLookupType.DONOR_IDS;
-
-    val donorSet1 = newArrayList("DO1", "DO3", "DO5", "DO7", "DO9");
-    termsLookupRepository.createTermsLookup(lookupType, id1, donorSet1);
-
-    val donorSet2 = newArrayList("DO2", "DO4", "DO5", "DO6", "DO8");
-    termsLookupRepository.createTermsLookup(lookupType, id2, donorSet2);
   }
 
 }