+++ CHANGES.md
@@ -2,6 +2,31 @@ ICGC DCC - Portal - Change Log
 ===
 
 Change log for the DCC Data Portal
+
+4.3.11
+--
+- Fixes regression for launching Survival Analysis demo. 
+- Don’t open manifest downloads in new tab, always download tarball
+- Now using HttpServletResponse to render manifests in order to eagerly commit status code and content disposition for user agents
+- Fixed issues with manifest generation when unique selected and using individually selected files
+- Fixed other UI issues related to manifest generation
+
+4.3.10
+--
+- Added ability to filter by compounds in Advanced Search
+- Added p-value calculations for Survival Analysis
+- Added ability to launch OncoGrid from Project page
+- Added ability in manifest dialog to include duplicates
+- Added support for `icgc-get` in Repository Browser
+- Changed Survival Analysis button to perform reset instead of undo
+- Fixed hidden close button in pathway viewer legend
+- Fixed text-wrapping issues
+- Fixed token manager to not error on unknown scopes
+
+4.3.9
+--
+- Fixed regression from Spring Boot upgrade: now ensures "_missing" count is present in response json, even when `null`
+
 4.3.8
 --
 - Upgraded to Spring Boot `1.4.0-RELEASE` and added custom `FailureAnalyzer`s for Elasticsearch, etc.
@@ -22,6 +47,8 @@ Change log for the DCC Data Portal
 
 4.3.5
 --
+  -  Made information table in pathway viewer sortable
+  -  Added mutation annotations to nodes, information table, and legend in pathway viewer
   -  Added exception mapper for 503 errors
   -  'Download unavailable' message while the service is down for maintenance 
   -  angular-animate and angular-cookies version update